feat(web): split Sessions and Traces into separate pages with vim-style tab nav
Promote Sessions and Traces to their own sidebar sections and routes
(/projects/$slug = Sessions, /projects/$slug/traces = Traces), extracting the
shared explorer shell into ProjectExplorer({ mode }) and removing the in-header
tab toggle. Add H/L tab navigation in the trace and session detail drawers, keep
J/K for item navigation (deferring to the span tree on Spans tabs), drop the
Alt+Arrow trace shortcuts, and stop the search bar from stealing focus on mount.
Repoint the tool/user "view traces/sessions" deep links to the new routes.
